refactor: move docx patcher to make for easier imports
