+++ src/pages-helpers/curriculum/docx/curriculum-download-patcher/patches/partnerDetail.ts
@@ -1,20 +1,20 @@
 import type { Element } from "xml-js";
 
-import { CombinedCurriculumData } from "../../../pages/teachers/curriculum/docx-poc/[...slugs]";
+import { CombinedCurriculumData } from "..";
 
 import { textIncludes, textReplacer } from "./util";
 
-export function partnerNamePatch(
+export function partnerDetailPatch(
   combinedCurriculumData: CombinedCurriculumData,
 ) {
   return async (el: Element) => {
-    if (el.type === "text" && textIncludes(el.text, "{{=PARTNER_NAME}}")) {
+    if (el.type === "text" && textIncludes(el.text, "{{=PARTNER_DETAIL}}")) {
       return {
         type: "text",
         text: textReplacer(
           el.text,
-          "{{=PARTNER_NAME}}",
-          combinedCurriculumData.curriculumPartner.name,
+          "{{=PARTNER_DETAIL}}",
+          combinedCurriculumData.partnerBio,
         ),
       };
     }
