+++ src/pages-helpers/curriculum/docx/curriculum-download-patcher/patches/partnerName.ts
@@ -1,20 +1,20 @@
 import type { Element } from "xml-js";
 
-import { CombinedCurriculumData } from "../../../pages/teachers/curriculum/docx-poc/[...slugs]";
+import { CombinedCurriculumData } from "..";
 
 import { textIncludes, textReplacer } from "./util";
 
-export function partnerDetailPatch(
+export function partnerNamePatch(
   combinedCurriculumData: CombinedCurriculumData,
 ) {
   return async (el: Element) => {
-    if (el.type === "text" && textIncludes(el.text, "{{=PARTNER_DETAIL}}")) {
+    if (el.type === "text" && textIncludes(el.text, "{{=PARTNER_NAME}}")) {
       return {
         type: "text",
         text: textReplacer(
           el.text,
-          "{{=PARTNER_DETAIL}}",
-          combinedCurriculumData.partnerBio,
+          "{{=PARTNER_NAME}}",
+          combinedCurriculumData.curriculumPartner.name,
         ),
       };
     }
