+++ src/pages-helpers/curriculum/docx/curriculum-download-patcher/patches/tableOfContents.ts
