+++ src/pages-helpers/curriculum/docx/curriculum-download-patcher/patches/unitNext.ts
@@ -1,6 +1,6 @@
 import type { Element } from "xml-js";
 
-import { CombinedCurriculumData } from "../../../pages/teachers/curriculum/docx-poc/[...slugs]";
+import { CombinedCurriculumData } from "..";
 
 import { textIncludes, textReplacer } from "./util";
