+++ src/pages-helpers/curriculum/docx/curriculum-download-patcher/patches/util.ts
