+++ src/pages/teachers/curriculum/docx-poc/[...slugs]/index.tsx
@@ -1,25 +1,5 @@
 import { OakThemeProvider, oakDefaultTheme } from "@oaknational/oak-components";
-import { Element } from "xml-js";
 
-import { unitsTablePatch } from "@/pages-helpers/curriculum/docx-patches/unitsTable";
-import { subjectPatch } from "@/pages-helpers/curriculum/docx-patches/subject";
-import { tableOfContentsPatch } from "@/pages-helpers/curriculum/docx-patches/tableOfContents";
-import { subjectExplainerPatch } from "@/pages-helpers/curriculum/docx-patches/subjectExplainer";
-import { partnerDetailPatch } from "@/pages-helpers/curriculum/docx-patches/partnerDetail";
-import { partnerNamePatch } from "@/pages-helpers/curriculum/docx-patches/partnerName";
-import { yearPatch } from "@/pages-helpers/curriculum/docx-patches/year";
-import { endOfDocumentPatch } from "@/pages-helpers/curriculum/docx-patches/endOfDocument";
-import { threadsTablePatch } from "@/pages-helpers/curriculum/docx-patches/threadsTable";
-import { unitTitlePatch } from "@/pages-helpers/curriculum/docx-patches/unitTitle";
-import { unitYearPatch } from "@/pages-helpers/curriculum/docx-patches/unitYear";
-import { unitLessonsPatch } from "@/pages-helpers/curriculum/docx-patches/unitLessons";
-import { mainThreadsPatch } from "@/pages-helpers/curriculum/docx-patches/mainThreads";
-import {
-  mapOverElements,
-  modifyXmlByRootSelector,
-  pipeElementThrough,
-  withBlock,
-} from "@/components/CurriculumComponents/DocxPOC/docx";
 import LiveDataSection from "@/components/CurriculumComponents/DocxPOC/components/LiveDataSection";
 import {
   download,
@@ -33,14 +13,9 @@ import curriculumApi2023, {
 } from "@/node-lib/curriculum-api-2023";
 import AppLayout from "@/components/SharedComponents/AppLayout";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
-import { unitThreadsPatch } from "@/pages-helpers/curriculum/docx-patches/unitThreads";
-import { unitPreviousPatch } from "@/pages-helpers/curriculum/docx-patches/unitPrevious";
-import { unitNextPatch } from "@/pages-helpers/curriculum/docx-patches/unitNext";
-import { unitNumberPatch } from "@/pages-helpers/curriculum/docx-patches/unitNumber";
-
-export type CombinedCurriculumData = CurriculumOverviewMVData &
-  CurriculumOverviewSanityData &
-  CurriculumUnitsTabData;
+import CurriculumDownlodsPatch, {
+  CombinedCurriculumData,
+} from "@/pages-helpers/curriculum/docx/curriculum-download-patcher";
 
 type PageProps = {
   combinedCurriculumData: CombinedCurriculumData;
@@ -55,8 +30,6 @@ export default function Page({
     examboardSlug ? ` - ${examboardSlug.toLocaleUpperCase()}` : ""
   }`;
 
-  // console.log({ combinedCurriculumData });
-
   const handleLiveDataClick = (file: File) => {
     const reader = new FileReader();
 
@@ -66,68 +39,9 @@ export default function Page({
       }
       const fileContent = event.target.result as ArrayBuffer;
       const uint8Array = new Uint8Array(fileContent);
-
-      // NOTE: This is really inefficient at the moment, that's fine for now though
-      const moddedFile = await modifyXmlByRootSelector(
+      const moddedFile = await CurriculumDownlodsPatch(
         uint8Array,
-        "w:document",
-        async (doc) => {
-          const docMod1 = await mapOverElements(
-            doc,
-            (el: Element, parent?: Element) => {
-              return pipeElementThrough(el, parent, [
-                unitsTablePatch(combinedCurriculumData),
-                subjectPatch(combinedCurriculumData),
-                tableOfContentsPatch(),
-                subjectExplainerPatch(combinedCurriculumData),
-                partnerDetailPatch(combinedCurriculumData),
-                partnerNamePatch(combinedCurriculumData),
-                yearPatch(),
-                threadsTablePatch(combinedCurriculumData),
-                endOfDocumentPatch(),
-              ]);
-            },
-          );
-
-          const docMod2 = await withBlock(
-            docMod1!,
-            "UNIT_PAGE",
-            async (el: Element) => {
-              const promises = combinedCurriculumData.units.map(
-                (unit, index) => {
-                  return mapOverElements(el, (el, parent) => {
-                    return pipeElementThrough(el, parent, [
-                      unitTitlePatch(unit),
-                      unitNumberPatch(unit, index),
-                      unitLessonsPatch(unit),
-                      unitYearPatch(unit),
-                      unitThreadsPatch(unit),
-                      unitPreviousPatch(unit),
-                      unitNextPatch(unit),
-                    ]);
-                  });
-                },
-              );
-              return await Promise.all(promises);
-            },
-          );
-
-          const docMod3 = await withBlock(
-            docMod2!,
-            "THREAD_PAGE",
-            async (el: Element, parent?: Element) => {
-              return pipeElementThrough(el, parent, [
-                mainThreadsPatch(combinedCurriculumData),
-              ]);
-            },
-          );
-
-          if (!docMod3) {
-            throw new Error("Invalid document!");
-          }
-
-          return docMod3;
-        },
+        combinedCurriculumData,
       );
 
       download(moddedFile, `${pageTitle} - ${formattedDate(new Date())}.docx`);
